+++ app/lib/config/index.js
@@ -2,7 +2,7 @@
 const dotenv = require("dotenv");
 dotenv.config();
 module.exports = {
-  healthPort: process.env.HEALTH_PORT || 9803,
+  healthPort: process.env.HEALTH_PORT || 9802,
   port: process.env.PORT || 8100,
   redisTimeout: 300,
   redis_host: process.env.REDIS_HOST || "redis-db"